backend/requirements.txt
fastapi
python-dotenv
uvicorn
emoji
SQLAlchemy
requests
requests-toolbelt
webencodings
zstandard
numpy
langchain
langchain-core
langchain-openai
langchain-text-splitters
langdetect
langsmith
lxml
nltk
beautifulsoup4
charset-normalizer
jsonpatch
jsonpointer
openai
PyYAML
pinecone-client
pinecone-plugin-interface
RapidFuzz
PyPDF2
docx
python-docx
python-multipart